backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
supabase==2.4.0
gotrue==2.9.1
postgrest==0.16.11
realtime==1.0.6
storage3==0.7.7
supafunc==0.3.3
python-dotenv==1.0.1
pydantic==2.6.3
websockets==12.0
python-multipart==0.0.9
httpx==0.25.2
httpcore==1.0.5